EMS-68 - restructure before-you-start structure and namings
