+++ ui/server/routes/index.js
@@ -1,11 +1,11 @@
 const express = require('express');
 const CONSTANTS = require('../../../constants');
-const getIndex = require('../controllers');
-const companyBased = require('../controllers/company-based');
+const beforeYouStartController = require('../controllers/before-you-start');
+const companyBasedController = require('../controllers/company-based');
 
 const router = express.Router();
 
-router.get(CONSTANTS.ROUTES.BEFORE_YOU_START, getIndex);
-router.get(CONSTANTS.ROUTES.COMPANY_BASED, companyBased);
+router.get(CONSTANTS.ROUTES.BEFORE_YOU_START, beforeYouStartController);
+router.get(CONSTANTS.ROUTES.COMPANY_BASED, companyBasedController);
 
 module.exports = router;
