+++ ui/server/routes/index.test.js
@@ -1,6 +1,6 @@
 const { get } = require('../test-mocks/mock-router');
 const CONSTANTS = require('../../../constants');
-const indexController = require('../controllers');
+const beforeYouStartController = require('../controllers/before-you-start');
 const companyBasedController = require('../controllers/company-based');
 
 describe('routes/index', () => {
@@ -15,7 +15,7 @@ describe('routes/index', () => {
   it('should setup all routes', () => {
     expect(get).toHaveBeenCalledTimes(2);
 
-    expect(get).toHaveBeenCalledWith(CONSTANTS.ROUTES.BEFORE_YOU_START, indexController);
+    expect(get).toHaveBeenCalledWith(CONSTANTS.ROUTES.BEFORE_YOU_START, beforeYouStartController);
     expect(get).toHaveBeenCalledWith(CONSTANTS.ROUTES.COMPANY_BASED, companyBasedController);
   });
 });
